backend-admin/requirements.txt
fastapi
uvicorn[standard]
motor
dnspython
PyJWT[cryptography]
python-dotenv
pydantic[email]
python-multipart
minio
bcrypt
requests>=2.31.0,<3
cryptography>=3.4.8
geopy>=2.3.0
bleach>=6.0.0
aiohttp>=3.8.0
cloudinary>=1.41.0,<2
Pillow>=10.0.0
python-dateutil>=2.8.0
